GEODE-8125: Extract AccessibleErrorCollector from SharedErrorCollector (#5112)
(cherry picked from commit f057b5c4dc8fa484e57987720605101a835acdaf)
